Cleaning the code
Implements DAO (Database Access Object) model to improve code cleaness

+++ app/controllers/AdminController.php
@@ -1,12 +1,19 @@
 <?php
 
-class AdminController extends \BaseController {
+class AdminController extends BaseController {
 
 	/**
 	 * Display a listing of the resource.
 	 *
 	 * @return Response
 	 */
+
+	protected $fileDao, $userDao;
+	public function __construct(FileDao $fileDao, UserDao $userDao){
+		$this->fileDao = $fileDao;
+		$this->userDao = $userDao;
+	}
+
 	public function index()
 	{
 		if(Auth::user()->is_admin){
@@ -26,7 +33,6 @@ class AdminController extends \BaseController {
 	public function create()
 	{
 		//
-		echo 'test';
 	}
 
 
@@ -52,15 +58,12 @@ class AdminController extends \BaseController {
 		//
 		if(Auth::user()->is_admin){
 			if($id=='files'){
-				$files = DB::table('files')
-							->join('users','users.id','=','files.id_user')
-							->select('users.username','files.id_user','files.path','files.key')
-							->get();
-				return View::make('admin.files')->with('files',$files);
+				return View::make('admin.files')
+					->with('files', $this->fileDao->getFilesAdmin());
 			}
 			else if($id=='users'){
-				$users = User::all();
-				return View::make('admin.users')->with('users',$users);	
+				return View::make('admin.users')
+					->with('users', $this->userDao->getUsers());	
 			}
 			else{
 				return Response::view('notfound');
@@ -72,11 +75,13 @@ class AdminController extends \BaseController {
 	}
 
 	public function files($id){
-		$files = UserFile::where('id_user','=',$id)->get();
-		$id_user = User::find($id);
-		$files->username=$id_user->username;
-		return View::make('admin.files')->with('files',$files);
+		$files = $this->fileDao->getFilesByOwnership($id);
+		$id_user = $this->userDao->getUserById($id);
+		$files->username = $id_user->username;
+		return View::make('admin.files')
+				->with('files', $files);
 	}
+
 	/**
 	 * Show the form for editing the specified resource.
 	 *
@@ -107,32 +112,40 @@ class AdminController extends \BaseController {
 	 * @param  int  $id
 	 * @return Response
 	 */
+	private function deleteFilesAndFolder($id){
+		$dir = public_path('files/'.$id);
+		if(file_exists($dir)){
+			$it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
+			$files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
+
+			foreach($files as $file) {
+			    if ($file->isDir()){
+					rmdir($file->getRealPath());
+				}
+				else{
+					unlink($file->getRealPath());
+				}
+			}
+			rmdir($dir);
+		}
+
+	}
+
 	public function destroy($id)
 	{
-		$dir = public_path('files/'.$id);
-		$it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
-		$files = new RecursiveIteratorIterator($it,
-					 RecursiveIteratorIterator::CHILD_FIRST);
-		foreach($files as $file) {
-		    if ($file->isDir()){
-				rmdir($file->getRealPath());
-			} else {
-				unlink($file->getRealPath());
-			}
-		}
-		rmdir($dir);
-		//unlink(public_path('files/2'));
-		UserFile::where('id_user','=',$id)->delete();
-		User::where('id','=',$id)->delete();
-		return Redirect::to('admin/users')->with('message','User and it\'s files succesfully deleted');
+		$this->deleteFilesAndFolder($id);
+		$this->fileDao->deleteFilesByOwnership($id);
+		$this->userDao->deleteUserById($id);
+		return Redirect::to('admin/users')
+				->with('message','User and it\'s files succesfully deleted');
 	}
 
 	public function delete($id)
 	{
 		if($id!=1){
-			$user = User::find($id);
-			$files = UserFile::where('id_user',$id)->get();
-			return View::make('admin.deleteUser')->with('user',$user)->with('files',$files);
+			return View::make('admin.deleteUser')
+				->with('user', $this->userDao->getUserById($id))
+				->with('files', $this->fileDao->getFilesByOwnership($id));
 		}
 		else{
 			return Response::view('unauthorized');
